Reorganizes folder structure; Renames sessionsComponent
