+++ src/app/landing-page/feedback/feedback.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core'
-import { environment } from '../../environments/environment'
+import { environment } from '../../../environments/environment'
 
 @Component({
   selector: 'app-feedback',
