+++ src/app/landing-page/schedule/schedule.component.css
