+++ src/app/partials/qr-code-scanner/qr-code-scanner.component.css
