+++ src/app/partials/qr-code-scanner/qr-code-scanner.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Output, EventEmitter } from '@angular/core'
-import { MessageService, Type } from '../partials/messages/message.service'
+import { MessageService, Type } from '../messages/message.service'
 
 @Component({
   selector: 'app-qrcode-scanner',
